requirements.txt
streamlit
speechrecognition
pyttsx3
gTTS
deep-translator
google-generativeai